+++ packages/kysely/src/sessions/remove.ts
@@ -4,9 +4,9 @@ import { Database } from "../db";
 export function remove(db: Kysely<Database>) {
   return async (tenant_id: string, id: string): Promise<boolean> => {
     const results = await db
-      .deleteFrom("sessions_2")
+      .deleteFrom("sessions")
       .where("tenant_id", "=", tenant_id)
-      .where("sessions_2.id", "=", id)
+      .where("sessions.id", "=", id)
       .execute();
 
     return !!results.length;